put flash initialization in board init for LPC55.  merged functions from board_flash.c into boards.c so that board_init can access the private variable _flash_config
