+++ src/main.c
@@ -131,11 +131,8 @@ int main(void)
   TU_LOG2("DFU init...\r\n");
   board_dfu_init();
 
-// Some platforms initialize flash earlier to check for a valid image
-#ifndef EARLY_FLASH_INIT  
   TU_LOG2("Flash init...\r\n");
   board_flash_init();
-#endif
 
   TU_LOG2("UF2 init...\r\n");
   uf2_init();
